Replace SQLite with Median Datastore for local data persistence
