+++ package.json
@@ -19,7 +19,6 @@
   },
   "dependencies": {
     "@catalyft/core": "file:packages/core",
-    "@capacitor-community/sqlite": "^7.0.1",
     "@capacitor/android": "^7.4.2",
     "@capacitor/cli": "^7.4.2",
     "@capacitor/core": "^7.4.2",
